web/requirements.txt
astroid==1.6.5
Django==2.0.7
django-split-settings==0.3.0
djangorestframework==3.8.2
flake8==3.5.0
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
psycopg2==2.7.5
psycopg2-binary==2.7.5
pycodestyle==2.3.1
pyflakes==1.6.0
pylint==1.9.2
pytz==2018.5
six==1.11.0
wrapt==1.10.11
coreapi==2.3.3
django-rest-swagger==2.2.0
